moved crop enums to enums.py
