+++ RUFAS/routines/feed_storage/hay.py
@@ -4,7 +4,7 @@ from RUFAS.general_constants import GeneralConstants
 from RUFAS.time import Time
 from RUFAS.weather import Weather
 
-from .enums import CropCategory
+from RUFAS.enums import CropCategory
 from .storage import Storage
 
 """
