+++ RUFAS/routines/field/crop/crop_configurations/corn.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from RUFAS.routines.feed_storage.enums import CropCategory, CropType
+from RUFAS.enums import CropCategory, CropType
 from RUFAS.routines.feed_storage.feed_manager import StorageType
 from RUFAS.routines.field.crop.crop_data import CropData, PlantCategory
 from RUFAS.routines.field.crop.crop_enum import CropSpecies
